fix: move App to components
