+++ src/components/App/App.js
@@ -1,17 +1,17 @@
 import React from 'react';
 import { Routes, Route } from 'react-router-dom';
-import './vendor/normalize.css';
-import './vendor/fonts/fonts.css';
+import '../../vendor/normalize.css';
+import '../../vendor/fonts/fonts.css';
 import './App.scss';
 
-import Header from './components/Header/Header';
-import Main from './pages/Main/Main';
-import Movies from './pages/Movies/Movies';
+import Header from '../Header/Header';
+import Main from '../../pages/Main/Main';
+import Movies from '../../pages/Movies/Movies';
 
 
 function App() {
   return (
-    <div className="page">
+    <div className="App">
       <Header />
       <main className="content">
         <Routes>
