+++ src/components/App/App.scss
@@ -1,6 +1,6 @@
-@import './utils/variables.scss';
+@import '../../utils/variables.scss';
 
-.page {
+.App {
   height: 100%;
   background-color: var(--color-light-black);
   font-family: Inter, Arial, sans-serif;
