+++ src/components/Promo/Promo.scss
@@ -1,4 +1,5 @@
 @import '../../utils/mixins.scss';
+@import '../../utils/variables.scss';
 
 .promo {
   background-color: var(--color-dark-blue);
